requirements.txt
absl-py==1.0.0
apiview==1.3.25
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-crispy-forms==1.14.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
executing==0.8.3
flatbuffers==2.0
fonttools==4.33.2
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.18.0
importlib-metadata==4.11.3
ipython==8.2.0
jedi==0.18.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.6
MarkupPy==1.14
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.22.3
oauthlib==3.2.0
odfpy==1.4.1
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
protobuf==3.20.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tablib==3.2.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-examples===7f254b19b9a651c99bf78c4921ca664885a2feae-
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
traitlets==5.1.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
